move compression header to compression.h
